split varibles
