+++ symbolic_trace/opcode_translator/executor/variables/__init__.py
@@ -1,1 +1,6 @@
 from .base import *  # noqa F401
+from .callable import *  # noqa F401
+from .container import *  # noqa F401
+from .iter import *  # noqa F401
+
+# TODO(zrr1999): import determinant
