+++ symbolic_trace/opcode_translator/executor/variables/base.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import collections
 import inspect
 import types
 from queue import Queue
@@ -10,24 +9,11 @@ import paddle
 
 from ....infer_meta import MetaInfo
 from ....symbolic.statement_ir import Symbol
-from ....utils import (
-    ASSERT,
-    NameGenerator,
-    is_break_graph_api,
-    is_paddle_api,
-    log_do,
-    paddle_tensor_methods,
-)
-from ....utils.exceptions import BreakGraphError, FallbackErrorBase, InnerError
+from ....utils import NameGenerator, log_do, paddle_tensor_methods
+from ....utils.exceptions import InnerError
 from ..guard import StringifyExpression, union_free_vars
 from ..pycode_generator import PyCodeGen
-from ..tracker import (
-    ConstTracker,
-    DummyTracker,
-    GetAttrTracker,
-    GetItemTracker,
-    Tracker,
-)
+from ..tracker import ConstTracker, DummyTracker, GetAttrTracker, Tracker
 
 if TYPE_CHECKING:
     from ..function_graph import FunctionGraph
@@ -81,6 +67,8 @@ def map_variables(map_func, variables):
         assert isinstance(
             variable, VariableBase
         ), f"variable must be VariableBase, got {variable}"
+        from .container import ContainerVariable
+
         if isinstance(variable, ContainerVariable):
             return paddle.utils.map_structure(
                 _map_variable, variable.get_wrapped_items()
@@ -93,13 +81,16 @@ def map_variables(map_func, variables):
 class VariableFactory:
     registered_funcs: list[Callable] = []
     mapping_priority_index: dict[int, int] = {}
+    default_priority = 0  # TODO(zrr1999): priority may be same in same file.
 
     @staticmethod
     def default_from_value(value, graph, tracker):
         return ObjectVariable(value, graph, tracker)
 
     @staticmethod
-    def register_from_value(priority: int = 0):
+    def register_from_value(priority: int | None = None):
+        if priority is None:
+            priority = VariableFactory.default_priority
         mapping_priority_index = VariableFactory.mapping_priority_index
         if mapping_priority_index.get(priority, None) is None:
             index = 0
@@ -177,6 +168,8 @@ class VariableBase:
         raise NotImplementedError()
 
     def flatten_items(self) -> list[VariableBase]:
+        from .container import ContainerVariable
+
         if not isinstance(self, ContainerVariable):
             return [self]
         flattened_items = []
@@ -214,6 +207,8 @@ class VariableBase:
             )
         attr = getattr(self.value, name)
         if inspect.ismethod(attr):
+            from .callable import UserDefinedMethodVariable
+
             return UserDefinedMethodVariable(
                 self,
                 attr.__func__,
@@ -227,7 +222,7 @@ class VariableBase:
     def getitem(self, *args, **kwargs):
         pass
 
-    @VariableFactory.register_from_value()
+    @VariableFactory.register_from_value(5)
     def from_value(
         value: Any,
         graph: FunctionGraph | None,
@@ -384,6 +379,8 @@ class TensorVariable(VariableBase):
 
     def __getattr__(self, name: str):
         if name in paddle_tensor_methods:
+            from .callable import TensorMethodVariable
+
             return TensorMethodVariable(
                 self, name, self.graph, tracker=GetAttrTracker(self, name)
             )
@@ -406,660 +403,14 @@ class TensorVariable(VariableBase):
         return None
 
 
-class ContainerVariable(VariableBase):
-    def get_items(self) -> list[VariableBase]:
-        raise NotImplementedError()
-
-    def __len__(self):
-        raise NotImplementedError()
-
-    def __bool__(self):
-        return len(self) > 0
-
-
-class ListVariable(ContainerVariable):
-    def __init__(
-        self,
-        val_list: list[VariableBase],
-        graph: FunctionGraph,
-        tracker: Tracker,
-    ):
+class ObjectVariable(VariableBase):
+    def __init__(self, obj, graph, tracker):
         super().__init__(tracker)
-        self.graph = graph
-        # everything in stack is VariableBase, so just accept the input list is ok
-        self.value = val_list
-
-    def get_value(self):
-        return [self[i].get_value() for i in range(len(self))]
-
-    def _reconstruct(self, codegen: PyCodeGen):
-        size = len(self)
-        for idx in range(size):
-            self[idx].reconstruct(codegen)
-        codegen.gen_build_list(size)
-
-    def get_items(self):
-        size = len(self)
-        return [self[idx] for idx in range(size)]
-
-    def get_wrapped_items(self):
-        return self.get_items()
-
-    def __repr__(self) -> str:
-        return f"ListVariable(len={len(self)})"
-
-    def __len__(self):
-        return len(self.value)
-
-    def __getitem__(self, key):
-        '''
-        we need to make sure that:
-            before an inplace change happens to ListVariable,
-            the related items should already be wrapped as VariableBase
-
-        if not, tracker might be set to a wrong elem
-        '''
-        if isinstance(key, VariableBase):
-            raise InnerError(
-                f"[{self.__class__.__name__}]: recieved {key} as key."
-            )
-
-        retval = self.value[key]
-
-        # if list is an input of funciton, we need make sure __getitem__ returns a VariableBase
-        retval = VariableFactory.from_value(
-            retval, self.graph, tracker=GetItemTracker(self, key)
-        )
-
-        return retval
-
-    def __setitem__(self, key, value):
-        '''
-        why __setitem__ is ok:
-
-        case:
-            def f(x = [t0, t1])
-                ...
-                x[0] = 0
-                ...
-
-            1. if setitem happens after get t0: t0 is a VariableBase (transformed at getitem), so it is ok
-            2. if setitem happens before get t0: t0 will not be used
-        '''
-        if isinstance(key, VariableBase):
-            raise InnerError(
-                f"[{self.__class__.__name__}]: received {key} as key."
-            )
-
-        if not isinstance(value, VariableBase):
-            raise InnerError(
-                f"[{self.__class__.__name__}]: received {value} to set value."
-            )
-        self.value[key] = value
-
-    def __delitem__(self, key):
-        if isinstance(key, VariableBase):
-            raise InnerError(
-                f"[{self.__class__.__name__}]: received {key} as key to delete."
-            )
-        del self.value[key]
-
-    @VariableFactory.register_from_value()
-    def from_value(value: Any, graph: FunctionGraph | None, tracker: Tracker):
-        if isinstance(value, list):
-            assert graph is not None
-            return ListVariable(value, graph=graph, tracker=tracker)
-        return None
-
-
-class TupleVariable(ContainerVariable):
-    def __init__(
-        self,
-        val_tuple: list[VariableBase],
-        graph: FunctionGraph,
-        tracker: Tracker,
-    ):
-        super().__init__(tracker)
-        self.graph = graph
-        # exactly it is a list (need replace item with VariableBase)
-        self.value = list(val_tuple)
-
-    def get_value(self):
-        return tuple(self[i].get_value() for i in range(len(self)))
-
-    def _reconstruct(self, codegen: PyCodeGen):
-        size = len(self)
-        for idx in range(size):
-            self[idx].reconstruct(codegen)
-        codegen.gen_build_tuple(size)
-
-    def get_items(self):
-        size = len(self)
-        return [self[idx] for idx in range(size)]
-
-    def get_wrapped_items(self):
-        return self.get_items()
-
-    def __repr__(self) -> str:
-        return f"TupleVariable(len={len(self)})"
-
-    def __len__(self):
-        return len(self.value)
-
-    def __getitem__(self, key):
-        if isinstance(key, VariableBase):
-            raise InnerError(
-                f"[{self.__class__.__name__}]: recieved {key} as key."
-            )
-        retval = self.value[key]
-
-        return VariableFactory.from_value(
-            retval, graph=self.graph, tracker=GetItemTracker(self, key)
-        )
-
-    def __setitem__(self, key, value):
-        raise InnerError(
-            f"[{self.__class__.__name__}]: setitem is not allowed."
-        )
-
-    def __delitem__(self, key):
-        raise InnerError(
-            f"[{self.__class__.__name__}]: delitem is not allowed."
-        )
-
-    @VariableFactory.register_from_value()
-    def from_value(value: Any, graph: FunctionGraph | None, tracker: Tracker):
-        if isinstance(value, tuple):
-            return TupleVariable(value, graph, tracker)
-        return None
-
-
-class DictVariable(ContainerVariable):
-    def __init__(
-        self,
-        val_dict: dict[object, VariableBase],
-        graph: FunctionGraph,
-        tracker: Tracker,
-    ):
-        super().__init__(tracker)
-        self.graph = graph
-        self.value = val_dict
-
-    def get_value(self):
-        return {key: self[key].get_value() for key in self.value}
-
-    def _reconstruct(self, codegen: PyCodeGen):
-        size = len(self)
-        for key in self.value.keys():
-            if not isinstance(key, ConstTypes):
-                raise InnerError(
-                    f"[{self.__class__.__name__}]: recieved {key} as key."
-                )
-            key_var = ConstantVariable.wrap_literal(key)
-            value_var = self[key]
-            key_var.reconstruct(codegen)
-            value_var.reconstruct(codegen)
-        codegen.gen_build_map(size)
-
-    def get_items(self):
-        items = []
-        for key in self.value.keys():
-            if not isinstance(key, ConstTypes):
-                raise InnerError(
-                    f"[{self.__class__.__name__}]: recieved {key} as key."
-                )
-            key_var = VariableFactory.from_value(
-                key, self.graph, tracker=ConstTracker(key)
-            )
-            value_var = self[key]
-            items.extend([key_var, value_var])
-        return items
-
-    def get_wrapped_items(self):
-        items = {}
-        for key in self.value.keys():
-            if not isinstance(key, ConstTypes):
-                raise InnerError(
-                    f"[{self.__class__.__name__}]: recieved {key} as key."
-                )
-            items[key] = self[key]
-        return items
-
-    def __repr__(self) -> str:
-        return f"DictVariable(len={len(self)})"
-
-    def __len__(self):
-        return len(self.value)
-
-    def __getitem__(self, key):
-        if isinstance(key, VariableBase):
-            raise InnerError(
-                f"[{self.__class__.__name__}]: recieved {key} as key."
-            )
-
-        retval = self.value[key]
-
-        return VariableFactory.from_value(
-            retval, self.graph, tracker=GetItemTracker(self, key)
-        )
-
-    def __setitem__(self, key, value):
-        if isinstance(key, VariableBase):
-            raise InnerError(
-                f"[{self.__class__.__name__}]: recieved {key} as key."
-            )
-
-        if not isinstance(value, ConstantVariable):
-            raise InnerError(
-                f"[{self.__class__.__name__}]: recieved {value} to set value."
-            )
-
-        self.value[key] = value
-
-    def __delitem__(self, key):
-        if isinstance(key, VariableBase):
-            raise InnerError(
-                f"[{self.__class__.__name__}]: recieved {key} as key to delete."
-            )
-        del self.value[key]
-
-    def override_method_keys(self):
-        raw_list = [
-            ConstantVariable(x, ConstTracker(x)) for x in self.value.keys()
-        ]
-        key_list = VariableFactory.from_value(
-            raw_list, self.graph, ConstTracker(raw_list)
-        )
-        return SequenceIterVariable(
-            key_list, self.graph, DummyTracker([key_list])
-        )
-
-    def override_method_values(self):
-        raw_list = list(self.get_wrapped_items().values())
-        value_list = VariableFactory.from_value(
-            raw_list, self.graph, DummyTracker([self])
-        )
-        return SequenceIterVariable(
-            value_list, self.graph, DummyTracker([value_list])
-        )
-
-    def override_method_items(self):
-        keys = [ConstantVariable(x, ConstTracker(x)) for x in self.value.keys()]
-        values = list(self.get_wrapped_items().values())
-        raw_list = list(zip(keys, values))
-        item_list = VariableFactory.from_value(
-            raw_list, self.graph, DummyTracker([self])
-        )
-        return SequenceIterVariable(
-            item_list, self.graph, DummyTracker([item_list])
-        )
-
-    def __getattr__(self, name):
-        name_ = "override_method_" + name
-        if hasattr(self, name_):
-            method = getattr(self, name_)
-            return DirectlyCallMethodVariable(
-                self,
-                method.__func__,
-                self.graph,
-                GetAttrTracker(self, name),
-            )
-        else:
-            raise NotImplementedError(
-                f"attribute {name} for dict is not implemented"
-            )
-
-    @VariableFactory.register_from_value()
-    def from_value(value: Any, graph: FunctionGraph | None, tracker: Tracker):
-        if isinstance(value, dict):
-            assert graph is not None
-            return DictVariable(value, graph=graph, tracker=tracker)
-
-
-class CallableVariable(VariableBase):
-    def __init__(self, graph: FunctionGraph, tracker: Tracker):
-        super().__init__(tracker)
+        self.value = obj
         self.graph = graph
 
-    def __call__(self, *args, **kwargs) -> VariableBase:
-        return self.call_function(*args, **kwargs)
-
-    def call_function(self, *args, **kwargs):
-        raise NotImplementedError("call_function is not implemented.")
-
-
-class FunctionVariable(CallableVariable):
-    def __init__(
-        self, fn: Callable[..., Any], graph: FunctionGraph, tracker: Tracker
-    ):
-        super().__init__(graph, tracker)
-        self.value = fn
-
-    def get_value(self):
-        return self.value
-
-    def get_code(self) -> types.CodeType:
-        return self.value.__code__
-
-
-class PaddleApiVariable(FunctionVariable):
-    def __init__(
-        self, fn: Callable[..., Any], graph: FunctionGraph, tracker: Tracker
-    ):
-        super().__init__(fn, graph, tracker)
-
-    def call_function(self, *args, **kwargs):
-        return self.graph.call_paddle_api(self.value, *args, **kwargs)
-
-    @VariableFactory.register_from_value()
-    def from_value(value: Any, graph: FunctionGraph | None, tracker: Tracker):
-        # This should be front of FunctionVariable to avoid conflict.
-        if callable(value) and is_paddle_api(value):
-            return PaddleApiVariable(value, graph, tracker)
-        return None
-
     def __repr__(self) -> str:
-        return f"PaddleApiVariable({self.value.__name__})"
-
-
-class UserDefinedGeneratorVariable(FunctionVariable):
-    def __init__(
-        self, fn: Callable[..., Any], graph: FunctionGraph, tracker: Tracker
-    ):
-        super().__init__(fn, graph, tracker)
-
-    def call_function(self, *args, **kwargs) -> VariableBase:
-
-        iter_ = self.value()
-        return VariableFactory.from_value(
-            iter_, self.graph, DummyTracker([self])
-        )
-
-    @VariableFactory.register_from_value()
-    def from_value(value: Any, graph: FunctionGraph | None, tracker: Tracker):
-        if inspect.isgeneratorfunction(value):
-            return UserDefinedGeneratorVariable(value, graph, tracker)
-        return None
-
-    def __repr__(self) -> str:
-        return f"UserDefinedGeneratorVariable({self.value.__name__})"
-
-
-class UserDefinedFunctionVariable(FunctionVariable):
-    def __init__(
-        self, fn: Callable[..., Any], graph: FunctionGraph, tracker: Tracker
-    ):
-        super().__init__(fn, graph, tracker)
-
-    def call_function(self, *args, **kwargs) -> VariableBase:
-        from ..opcode_inline_executor import OpcodeInlineExecutor
-
-        if self.value is ASSERT:
-            return self.value(args[0].value)
-
-        checkpoint = self.graph.save_memo()
-        try:
-            inline_executor = OpcodeInlineExecutor(self, *args, **kwargs)
-            output = inline_executor.inline_call()
-        except FallbackErrorBase as e:
-            self.graph.restore_memo(checkpoint)
-            raise BreakGraphError(
-                f"{self.value} is raise a inline call error. {e}"
-            )
-        return output
-
-    @VariableFactory.register_from_value()
-    def from_value(value: Any, graph: FunctionGraph | None, tracker: Tracker):
-        if isinstance(value, (types.FunctionType)):
-            return UserDefinedFunctionVariable(value, graph, tracker)
-        return None
-
-    def __repr__(self) -> str:
-        return f"UserDefinedFunctionVariable({self.value.__name__})"
-
-
-class MethodVariable(CallableVariable):
-    def __init__(
-        self,
-        bound_instance: VariableBase,
-        graph: FunctionGraph,
-        tracker: Tracker,
-    ):
-        super().__init__(graph, tracker)
-        self.bound_instance = bound_instance
-
-
-class TensorMethodVariable(MethodVariable):
-    def __init__(
-        self,
-        tensor: TensorVariable,
-        method_name: str,
-        graph: FunctionGraph,
-        tracker: Tracker,
-    ):
-        super().__init__(tensor, graph, tracker)
-        self.tensor = tensor
-        self.method_name = method_name
-
-    def get_value(self):
-        return getattr(self.tensor, self.method_name)
-
-    def call_function(self, *args, **kwargs):
-        return self.graph.call_tensor_method(
-            self.method_name, self.tensor, *args, **kwargs
-        )
-
-    @VariableFactory.register_from_value()
-    def from_value(value: Any, graph: FunctionGraph | None, tracker: Tracker):
-        if inspect.ismethod(value) and isinstance(
-            value.__self__, paddle.Tensor
-        ):
-            # NOTE(SigureMo): Since the method_self need method_var as the obj
-            # of the tracker, we need to temporarily set the tracker of method_self
-            # to DummyTracker, and set it to GetAttrTracker after method_var is created.
-            method_self = TensorVariable(
-                value.__self__, graph, DummyTracker([])
-            )
-            method_var = TensorMethodVariable(
-                method_self,
-                value.__name__,
-                graph,
-                tracker,
-            )
-            method_self.tracker = GetAttrTracker(method_var, "__self__")
-            return method_var
-        return None
-
-    def __repr__(self) -> str:
-        return f"TensorMethodVariable({self.method_name})"
-
-
-class UserDefinedMethodVariable(MethodVariable):
-    def __init__(
-        self, bound_instance, fn, graph: FunctionGraph, tracker: Tracker
-    ):
-        super().__init__(bound_instance, graph, tracker)
-        self.bound_instance = bound_instance
-        self.fn = fn
-
-    def get_value(self):
-        return self.fn.__get__(
-            self.bound_instance, self.bound_instance.__class__
-        )
-
-    def call_function(self, *args, **kwargs):
-        fn_var = UserDefinedFunctionVariable(
-            self.fn, self.graph, GetAttrTracker(self, "__func__")
-        )
-
-        return fn_var(*(self.bound_instance, *args), **kwargs)
-
-    @VariableFactory.register_from_value()
-    def from_value(value: Any, graph: FunctionGraph | None, tracker: Tracker):
-        if inspect.ismethod(value):
-            method_self = VariableFactory.from_value(
-                value.__self__, graph, DummyTracker([])
-            )
-            method_var = UserDefinedMethodVariable(
-                method_self,
-                value.__func__,
-                graph,
-                tracker,
-            )
-            method_self.tracker = GetAttrTracker(method_var, "__self__")
-            return method_var
-        return None
-
-    def __repr__(self) -> str:
-        return f"UserDefinedMethodVariable({self.fn.__name__})"
-
-
-class DirectlyCallMethodVariable(MethodVariable):
-    def __init__(
-        self, bound_instance, fn, graph: FunctionGraph, tracker: Tracker
-    ):
-        super().__init__(bound_instance, graph, tracker)
-        self.bound_instance = bound_instance
-        self.fn = fn
-
-    def get_value(self):
-        return self.fn.__get__(
-            self.bound_instance, self.bound_instance.__class__
-        )
-
-    def call_function(self, *args, **kwargs):
-        return self.fn(*(self.bound_instance, *args), **kwargs)
-
-
-class LayerVariable(CallableVariable):
-    def __init__(
-        self, layer: paddle.nn.Layer, graph: FunctionGraph, tracker: Tracker
-    ):
-        super().__init__(graph, tracker)
-        self.value = layer
-
-    def get_value(self):
-        return self.value
-
-    def make_stringify_guard(self) -> StringifyExpression:
-        assert not isinstance(
-            self.tracker, DummyTracker
-        ), "Can not make guard from dummy tracker"
-
-        frame_value_tracer = self.tracker.trace_value_from_frame()
-        log_do(
-            4,
-            lambda: print(
-                f"[Guard]: guard_fn for {self}, tracker={self.tracker.__class__.__name__}, value={frame_value_tracer.expr}"
-            ),
-        )
-        return StringifyExpression(
-            f"id({frame_value_tracer.expr}) == {id(self.get_value())}",
-            union_free_vars(frame_value_tracer.free_vars),
-        ) & StringifyExpression(
-            f"{frame_value_tracer.expr}.training == {self.get_value().training}",
-            union_free_vars(frame_value_tracer.free_vars),
-        )
-
-
-class PaddleLayerVariable(LayerVariable):
-    layer_name_generator = NameGenerator("layer_")
-
-    def __init__(
-        self, layer: paddle.nn.Layer, graph: FunctionGraph, tracker: Tracker
-    ):
-        super().__init__(layer, graph, tracker)
-        self.name = self.layer_name_generator.next()
-
-    def get_symbol(self) -> Symbol:
-        return Symbol(self.name)
-
-    def call_function(self, *args, **kwargs):
-        # TODO: Remove this trick after we support for-loop.
-        if isinstance(self.value, paddle.nn.Sequential):
-            assert len(args) == 1, "Sequential only accept one input"
-            input = args[0]
-            for i, layer in enumerate(self.value._sub_layers.values()):
-                layer_var = VariableFactory.from_value(
-                    layer, self.graph, tracker=GetItemTracker(self, i)
-                )
-                assert isinstance(layer_var, LayerVariable)
-                input = layer_var(input)
-            return input
-        return self.graph.call_layer(self, *args, **kwargs)
-
-    @VariableFactory.register_from_value()
-    def from_value(value: Any, graph: FunctionGraph | None, tracker: Tracker):
-        # TODO(SigureMo): Add a more common way to check if a value is a paddle builtin layer.
-        if isinstance(value, paddle.nn.Layer) and value.__module__.startswith(
-            "paddle.nn."
-        ):
-            return PaddleLayerVariable(value, graph, tracker)
-        return None
-
-    def __repr__(self) -> str:
-        return f"PaddleLayerVariable({self.value.__class__.__name__})"
-
-
-class UserDefinedLayerVariable(LayerVariable):
-    def __init__(
-        self, layer: paddle.nn.Layer, graph: FunctionGraph, tracker: Tracker
-    ):
-        super().__init__(layer, graph, tracker)
-
-    def call_function(self, *args, **kwargs):
-        fn_var = UserDefinedFunctionVariable(
-            self.value.__class__.__call__,
-            self.graph,
-            GetAttrTracker(self, "__call__"),
-        )
-
-        return fn_var(*(self, *args), **kwargs)
-
-    @VariableFactory.register_from_value()
-    def from_value(value: Any, graph: FunctionGraph | None, tracker: Tracker):
-        if isinstance(
-            value, paddle.nn.Layer
-        ) and not value.__module__.startswith("paddle.nn."):
-            return UserDefinedLayerVariable(value, graph, tracker)
-        return None
-
-    def __repr__(self) -> str:
-        return f"UserDefinedLayerVariable({self.value.__class__.__name__})"
-
-
-class BuiltinVariable(CallableVariable):
-    def __init__(
-        self, fn: Callable[..., Any], graph: FunctionGraph, tracker: Tracker
-    ):
-        super().__init__(graph, tracker)
-        self.value = fn
-
-    def call_function(self, *args, **kwargs):
-        # TODO(0x45f): For builtin functions, may have 3 different ways to process as below:
-        #     1. Simulation execution: ensure correct simulation execution and handle trackers with care
-        #     2. Trigger the paddle api call
-        #     3. Trigger fallback
-        if is_break_graph_api(self.value):
-            raise BreakGraphError()
-        args = [
-            arg.value if isinstance(arg, ConstantVariable) else arg
-            for arg in args
-        ]
-        kwargs = {
-            k: (v.value if isinstance(v, ConstantVariable) else v)
-            for k, v in kwargs.items()
-        }
-        return self.value(*args, **kwargs)
-
-    @VariableFactory.register_from_value()
-    def from_value(value: Any, graph: FunctionGraph | None, tracker: Tracker):
-        if isinstance(value, (types.BuiltinFunctionType)):
-            return BuiltinVariable(value, graph, tracker)
-        return None
-
-    def __repr__(self) -> str:
-        return f"BuiltinVariable({self.value.__name__})"
+        return f"ObjectVariable({self.value})"
 
 
 class SliceVariable(VariableBase):
@@ -1132,67 +483,3 @@ class DygraphTracerVariable(VariableBase):
         if isinstance(value, paddle.fluid.dygraph.tracer.Tracer):
             return DygraphTracerVariable(value, graph, tracker)
         return None
-
-
-class ObjectVariable(VariableBase):
-    def __init__(self, obj, graph, tracker):
-        super().__init__(tracker)
-        self.value = obj
-        self.graph = graph
-
-    def __repr__(self) -> str:
-        return f"ObjectVariable({self.value})"
-
-
-class IterVariable(VariableBase):
-    def __init__(self, obj, graph, tracker):
-        super().__init__(tracker)
-        self.hold = obj
-        self.graph = graph
-
-    @VariableFactory.register_from_value()
-    def from_value(value: Any, graph: FunctionGraph | None, tracker: Tracker):
-        if isinstance(value, collections.abc.Iterable):
-            return UserDefinedIterVariable(value, graph, tracker)
-        return None
-
-
-class SequenceIterVariable(IterVariable):
-    def __init__(self, obj, graph, tracker):
-        super().__init__(obj, graph, tracker)
-        self.idx = 0
-
-    def next(self):
-        if self.idx < len(self.hold):
-            val = self.hold[self.idx]
-            self.idx += 1
-            return val
-        else:
-            raise StopIteration()
-
-
-class DictIterVariable(IterVariable):
-    def __init__(self, obj, graph, tracker):
-        super().__init__(obj, graph, tracker)
-        self.key_list = [
-            ConstantVariable(x, ConstTracker(x)) for x in self.hold
-        ]
-        self.idx = 0
-
-    def next(self):
-        if self.idx < len(self.key_list):
-            val = self.key_list[self.idx]
-            return val
-        else:
-            raise StopIteration()
-
-
-class TensorIterVariable(IterVariable):
-    def __init__(self, obj, graph, tracker):
-        super().__init__(obj, graph, tracker)
-
-
-# what UserDefinedIterVariable holds doesn't matter, because use user defined iterator will trigger break graph
-class UserDefinedIterVariable(IterVariable):
-    def __init__(self, obj, graph, tracker):
-        super().__init__(obj, graph, tracker)
