+++ symbolic_trace/opcode_translator/executor/variables/container.py
@@ -1,0 +1,337 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Any
+
+from ....utils.exceptions import InnerError
+from ..pycode_generator import PyCodeGen
+from ..tracker import (
+    ConstTracker,
+    DummyTracker,
+    GetAttrTracker,
+    GetItemTracker,
+    Tracker,
+)
+from .base import ConstantVariable, ConstTypes, VariableBase, VariableFactory
+
+if TYPE_CHECKING:
+    from ..function_graph import FunctionGraph
+
+
+class ContainerVariable(VariableBase):
+    def get_items(self) -> list[VariableBase]:
+        raise NotImplementedError()
+
+    def __len__(self):
+        raise NotImplementedError()
+
+    def __bool__(self):
+        return len(self) > 0
+
+
+class ListVariable(ContainerVariable):
+    def __init__(
+        self,
+        val_list: list[VariableBase],
+        graph: FunctionGraph,
+        tracker: Tracker,
+    ):
+        super().__init__(tracker)
+        self.graph = graph
+        # everything in stack is VariableBase, so just accept the input list is ok
+        self.value = val_list
+
+    def get_value(self):
+        return [self[i].get_value() for i in range(len(self))]
+
+    def _reconstruct(self, codegen: PyCodeGen):
+        size = len(self)
+        for idx in range(size):
+            self[idx].reconstruct(codegen)
+        codegen.gen_build_list(size)
+
+    def get_items(self):
+        size = len(self)
+        return [self[idx] for idx in range(size)]
+
+    def get_wrapped_items(self):
+        return self.get_items()
+
+    def __repr__(self) -> str:
+        return f"ListVariable(len={len(self)})"
+
+    def __len__(self):
+        return len(self.value)
+
+    def __getitem__(self, key):
+        '''
+        we need to make sure that:
+            before an inplace change happens to ListVariable,
+            the related items should already be wrapped as VariableBase
+
+        if not, tracker might be set to a wrong elem
+        '''
+        if isinstance(key, VariableBase):
+            raise InnerError(
+                f"[{self.__class__.__name__}]: recieved {key} as key."
+            )
+
+        retval = self.value[key]
+
+        # if list is an input of funciton, we need make sure __getitem__ returns a VariableBase
+        retval = VariableFactory.from_value(
+            retval, self.graph, tracker=GetItemTracker(self, key)
+        )
+
+        return retval
+
+    def __setitem__(self, key, value):
+        '''
+        why __setitem__ is ok:
+
+        case:
+            def f(x = [t0, t1])
+                ...
+                x[0] = 0
+                ...
+
+            1. if setitem happens after get t0: t0 is a VariableBase (transformed at getitem), so it is ok
+            2. if setitem happens before get t0: t0 will not be used
+        '''
+        if isinstance(key, VariableBase):
+            raise InnerError(
+                f"[{self.__class__.__name__}]: received {key} as key."
+            )
+
+        if not isinstance(value, VariableBase):
+            raise InnerError(
+                f"[{self.__class__.__name__}]: received {value} to set value."
+            )
+        self.value[key] = value
+
+    def __delitem__(self, key):
+        if isinstance(key, VariableBase):
+            raise InnerError(
+                f"[{self.__class__.__name__}]: received {key} as key to delete."
+            )
+        del self.value[key]
+
+    @VariableFactory.register_from_value()
+    def from_value(value: Any, graph: FunctionGraph | None, tracker: Tracker):
+        if isinstance(value, list):
+            assert graph is not None
+            return ListVariable(value, graph=graph, tracker=tracker)
+        return None
+
+
+class TupleVariable(ContainerVariable):
+    def __init__(
+        self,
+        val_tuple: list[VariableBase],
+        graph: FunctionGraph,
+        tracker: Tracker,
+    ):
+        super().__init__(tracker)
+        self.graph = graph
+        # exactly it is a list (need replace item with VariableBase)
+        self.value = list(val_tuple)
+
+    def get_value(self):
+        return tuple(self[i].get_value() for i in range(len(self)))
+
+    def _reconstruct(self, codegen: PyCodeGen):
+        size = len(self)
+        for idx in range(size):
+            self[idx].reconstruct(codegen)
+        codegen.gen_build_tuple(size)
+
+    def get_items(self):
+        size = len(self)
+        return [self[idx] for idx in range(size)]
+
+    def get_wrapped_items(self):
+        return self.get_items()
+
+    def __repr__(self) -> str:
+        return f"TupleVariable(len={len(self)})"
+
+    def __len__(self):
+        return len(self.value)
+
+    def __getitem__(self, key):
+        if isinstance(key, VariableBase):
+            raise InnerError(
+                f"[{self.__class__.__name__}]: recieved {key} as key."
+            )
+        retval = self.value[key]
+
+        return VariableFactory.from_value(
+            retval, graph=self.graph, tracker=GetItemTracker(self, key)
+        )
+
+    def __setitem__(self, key, value):
+        raise InnerError(
+            f"[{self.__class__.__name__}]: setitem is not allowed."
+        )
+
+    def __delitem__(self, key):
+        raise InnerError(
+            f"[{self.__class__.__name__}]: delitem is not allowed."
+        )
+
+    @VariableFactory.register_from_value()
+    def from_value(value: Any, graph: FunctionGraph | None, tracker: Tracker):
+        if isinstance(value, tuple):
+            return TupleVariable(value, graph, tracker)
+        return None
+
+
+class DictVariable(ContainerVariable):
+    def __init__(
+        self,
+        val_dict: dict[object, VariableBase],
+        graph: FunctionGraph,
+        tracker: Tracker,
+    ):
+        super().__init__(tracker)
+        self.graph = graph
+        self.value = val_dict
+
+    def get_value(self):
+        return {key: self[key].get_value() for key in self.value}
+
+    def _reconstruct(self, codegen: PyCodeGen):
+        from .base import ConstantVariable
+
+        size = len(self)
+        for key in self.value.keys():
+            if not isinstance(key, ConstTypes):
+                raise InnerError(
+                    f"[{self.__class__.__name__}]: recieved {key} as key."
+                )
+            key_var = ConstantVariable.wrap_literal(key)
+            value_var = self[key]
+            key_var.reconstruct(codegen)
+            value_var.reconstruct(codegen)
+        codegen.gen_build_map(size)
+
+    def get_items(self):
+        items = []
+        for key in self.value.keys():
+            if not isinstance(key, ConstTypes):
+                raise InnerError(
+                    f"[{self.__class__.__name__}]: recieved {key} as key."
+                )
+            key_var = VariableFactory.from_value(
+                key, self.graph, tracker=ConstTracker(key)
+            )
+            value_var = self[key]
+            items.extend([key_var, value_var])
+        return items
+
+    def get_wrapped_items(self):
+        items = {}
+        for key in self.value.keys():
+            if not isinstance(key, ConstTypes):
+                raise InnerError(
+                    f"[{self.__class__.__name__}]: recieved {key} as key."
+                )
+            items[key] = self[key]
+        return items
+
+    def __repr__(self) -> str:
+        return f"DictVariable(len={len(self)})"
+
+    def __len__(self):
+        return len(self.value)
+
+    def __getitem__(self, key):
+        if isinstance(key, VariableBase):
+            raise InnerError(
+                f"[{self.__class__.__name__}]: recieved {key} as key."
+            )
+
+        retval = self.value[key]
+
+        return VariableFactory.from_value(
+            retval, self.graph, tracker=GetItemTracker(self, key)
+        )
+
+    def __setitem__(self, key, value):
+        if isinstance(key, VariableBase):
+            raise InnerError(
+                f"[{self.__class__.__name__}]: recieved {key} as key."
+            )
+
+        if not isinstance(value, ConstantVariable):
+            raise InnerError(
+                f"[{self.__class__.__name__}]: recieved {value} to set value."
+            )
+
+        self.value[key] = value
+
+    def __delitem__(self, key):
+        if isinstance(key, VariableBase):
+            raise InnerError(
+                f"[{self.__class__.__name__}]: recieved {key} as key to delete."
+            )
+        del self.value[key]
+
+    def override_method_keys(self):
+        from .iter import SequenceIterVariable
+
+        raw_list = [
+            ConstantVariable(x, ConstTracker(x)) for x in self.value.keys()
+        ]
+        key_list = VariableFactory.from_value(
+            raw_list, self.graph, ConstTracker(raw_list)
+        )
+        return SequenceIterVariable(
+            key_list, self.graph, DummyTracker([key_list])
+        )
+
+    def override_method_values(self):
+        from .iter import SequenceIterVariable
+
+        raw_list = list(self.get_wrapped_items().values())
+        value_list = VariableFactory.from_value(
+            raw_list, self.graph, DummyTracker([self])
+        )
+        return SequenceIterVariable(
+            value_list, self.graph, DummyTracker([value_list])
+        )
+
+    def override_method_items(self):
+        from .iter import SequenceIterVariable
+
+        keys = [ConstantVariable(x, ConstTracker(x)) for x in self.value.keys()]
+        values = list(self.get_wrapped_items().values())
+        raw_list = list(zip(keys, values))
+        item_list = VariableFactory.from_value(
+            raw_list, self.graph, DummyTracker([self])
+        )
+        return SequenceIterVariable(
+            item_list, self.graph, DummyTracker([item_list])
+        )
+
+    def __getattr__(self, name):
+        from .callable import DirectlyCallMethodVariable
+
+        name_ = "override_method_" + name
+        if hasattr(self, name_):
+            method = getattr(self, name_)
+            return DirectlyCallMethodVariable(
+                self,
+                method.__func__,
+                self.graph,
+                GetAttrTracker(self, name),
+            )
+        else:
+            raise NotImplementedError(
+                f"attribute {name} for dict is not implemented"
+            )
+
+    @VariableFactory.register_from_value()
+    def from_value(value: Any, graph: FunctionGraph | None, tracker: Tracker):
+        if isinstance(value, dict):
+            assert graph is not None
+            return DictVariable(value, graph=graph, tracker=tracker)
