+++ symbolic_trace/opcode_translator/executor/variables/iter.py
@@ -1,0 +1,64 @@
+from __future__ import annotations
+
+import collections
+from typing import TYPE_CHECKING, Any
+
+from ..tracker import ConstTracker, Tracker
+from .base import ConstantVariable, VariableBase, VariableFactory
+
+if TYPE_CHECKING:
+    from ..function_graph import FunctionGraph
+
+
+class IterVariable(VariableBase):
+    def __init__(self, obj, graph, tracker):
+        super().__init__(tracker)
+        self.hold = obj
+        self.graph = graph
+
+    @VariableFactory.register_from_value()
+    def from_value(value: Any, graph: FunctionGraph | None, tracker: Tracker):
+        if isinstance(value, collections.abc.Iterable):
+            return UserDefinedIterVariable(value, graph, tracker)
+        return None
+
+
+class SequenceIterVariable(IterVariable):
+    def __init__(self, obj, graph, tracker):
+        super().__init__(obj, graph, tracker)
+        self.idx = 0
+
+    def next(self):
+        if self.idx < len(self.hold):
+            val = self.hold[self.idx]
+            self.idx += 1
+            return val
+        else:
+            raise StopIteration()
+
+
+class DictIterVariable(IterVariable):
+    def __init__(self, obj, graph, tracker):
+        super().__init__(obj, graph, tracker)
+        self.key_list = [
+            ConstantVariable(x, ConstTracker(x)) for x in self.hold
+        ]
+        self.idx = 0
+
+    def next(self):
+        if self.idx < len(self.key_list):
+            val = self.key_list[self.idx]
+            return val
+        else:
+            raise StopIteration()
+
+
+class TensorIterVariable(IterVariable):
+    def __init__(self, obj, graph, tracker):
+        super().__init__(obj, graph, tracker)
+
+
+# what UserDefinedIterVariable holds doesn't matter, because use user defined iterator will trigger break graph
+class UserDefinedIterVariable(IterVariable):
+    def __init__(self, obj, graph, tracker):
+        super().__init__(obj, graph, tracker)
